+++ apm-agent-core/src/main/java/co/elastic/apm/configuration/CoreConfiguration.java
@@ -36,7 +36,7 @@ public class CoreConfiguration extends ConfigurationOptionProvider {
     public static final String ACTIVE = "active";
     public static final String INSTRUMENT = "instrument";
     public static final String SERVICE_NAME = "service_name";
-    public static final String SAMPLE_RATE = "sample_rate";
+    public static final String SAMPLE_RATE = "transaction_sample_rate";
     private static final String CORE_CATEGORY = "Core";
     private final ConfigurationOption<Boolean> active = ConfigurationOption.booleanOption()
         .key(ACTIVE)
@@ -81,11 +81,17 @@ public class CoreConfiguration extends ConfigurationOptionProvider {
     private final ConfigurationOption<String> environment = ConfigurationOption.stringOption()
         .key("environment")
         .configurationCategory(CORE_CATEGORY)
-        .description("The name of the environment this service is deployed in, e.g. \"production\" or \"staging\".")
+        .description("The name of the environment this service is deployed in, e.g. \"production\" or \"staging\"." +
+            "" +
+            "NOTE: The APM UI does not fully support the environment setting yet.\n" +
+            "You can use the query bar to filter for a specific environment,\n" +
+            "but by default the environments will be mixed together.\n" +
+            "Also keep that in mind when creating alerts.")
         .build();
 
     private final ConfigurationOption<Double> sampleRate = ConfigurationOption.doubleOption()
         .key(SAMPLE_RATE)
+        .aliasKeys("sample_rate")
         .configurationCategory(CORE_CATEGORY)
         .description("By default, the agent will sample every transaction (e.g. request to your service). " +
             "To reduce overhead and storage requirements, you can set the sample rate to a value between 0.0 and 1.0. " +